Remove Enumerable from RichDocument for Sunspot 2.2.7/RSolr 2.0
Sunspot 2.2.7 upgraded to RSolr 2.0.
RSolr 2.0 added JSON support and set it as the default generator.
The JSON generator expects to call `as_json` on `RSolr::Xml::Document`.
However, since `RichDocument` was mixing-in Enumerable, the ActiveSupport implementation of `as_json` was replacing the method defined in `RSolr::Xml::Document`. This was breaking indexing for documents without attachments, since they are still `RichDocument`s.
This is only an issue for codebases that have ActiveSupport, which is the primary target of Sunspot.

I don't see why `Enumerable` is mixed into this class.
This class does not implement `#each`, therefore I don't understand how `Enumerable` could be useful here.